refactor: rename apim/apidefinition/delegate pkg to internal
