+++ controllers/apim/apidefinition/apidefinition_controller.go
@@ -30,7 +30,7 @@ import (
 
 	"github.com/go-logr/logr"
 	gio "github.com/gravitee-io/gravitee-kubernetes-operator/api/v1alpha1"
-	"github.com/gravitee-io/gravitee-kubernetes-operator/controllers/apim/apidefinition/delegate"
+	"github.com/gravitee-io/gravitee-kubernetes-operator/controllers/apim/apidefinition/internal"
 	"github.com/gravitee-io/gravitee-kubernetes-operator/controllers/apim/managementcontext"
 
 	"github.com/gravitee-io/gravitee-kubernetes-operator/pkg/keys"
@@ -83,7 +83,7 @@ func (r *Reconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Resu
 	log = log.WithValues("apiDefinitionName", apiDefinition.Name)
 
 	// Ini Delegate and set management context if defined
-	apisDelegate := delegate.NewDelegate(ctx, r.Client, log)
+	apisDelegate := internal.NewDelegate(ctx, r.Client, log)
 
 	if apiDefinition.Spec.Context != nil {
 		managementContext, ctxErr := managementcontext.Get(ctx, r.Client, log, apiDefinition.Spec.Context)
