+++ controllers/apim/apidefinition/internal/config_map.go
@@ -1,4 +1,4 @@
-package delegate
+package internal
 
 import (
 	"encoding/json"
