+++ controllers/apim/apidefinition/internal/create.go
@@ -1,4 +1,4 @@
-package delegate
+package internal
 
 import (
 	"encoding/json"
