+++ controllers/apim/apidefinition/internal/finalizer.go
@@ -1,4 +1,4 @@
-package delegate
+package internal
 
 import (
 	gio "github.com/gravitee-io/gravitee-kubernetes-operator/api/v1alpha1"
