+++ controllers/apim/apidefinition/internal/import_api_definition_template.go
@@ -1,4 +1,4 @@
-package delegate
+package internal
 
 import (
 	"fmt"
