+++ controllers/apim/apidefinition/internal/update.go
@@ -1,4 +1,4 @@
-package delegate
+package internal
 
 import (
 	"encoding/json"
