+++ controllers/apim/apidefinition/internal/update_api_state.go
@@ -1,4 +1,4 @@
-package delegate
+package internal
 
 import (
 	"fmt"
